test_requirements.txt
factory_boy==2.8.1
Faker==0.7.17
flake8==3.3.0
pre_commit==0.15.0
pytest==3.0.7
pytest-django==3.1.2
testing.postgresql==1.3.0
